aws cloudstorage bucket refactor (#561)
* aws cloudstorage bucket refactor

* remove unused configmap

* azure wip

wip

reset skeleton

update

wip

wip

* remove added internal methods from interface, passes more parameters

* indent

* remove wip azure.go

* address https://github.com/openshift/oadp-operator/pull/561#pullrequestreview-878255237
