+++ bundle/manifests/oadp.openshift.io_cloudstorages.yaml
@@ -55,7 +55,8 @@ spec:
                   for AWS Buckets
                 type: boolean
               name:
-                description: Name is the name requested for the bucket
+                description: Name is the name requested for the bucket (aws, gcp)
+                  or container (azure)
                 type: string
               provider:
                 enum:
