+++ config/crd/bases/oadp.openshift.io_cloudstorages.yaml
@@ -57,7 +57,8 @@ spec:
                   for AWS Buckets
                 type: boolean
               name:
-                description: Name is the name requested for the bucket
+                description: Name is the name requested for the bucket (aws, gcp)
+                  or container (azure)
                 type: string
               provider:
                 enum:
